Add files via upload
I developed a program that managed a database for a retail flower shop using object-oriented programming with at least three classes, where one of them (the driver) was called FlowersDriver. The other two were named Flower and DatabaseTotals. The main method of the program was in the class FlowersDriver and included a loop that allowed users to enter database commands, one at a time, from a menu. The exact number of classes used was up to my design discretion, but I used at least these three. Once a command was selected (done by typing in an integer on the Console), the program executed that command, and the user was prompted to select another command indefinitely until they chose the command that terminated the program.